Move custom query operators to query package
