+++ src/ggrc/query/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2017 Google Inc.
+# Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
